moved PageContainer to pages folder, updated login event handler using hisotry to push my account page on login (in progress)
